Switch to floating point representation of color
This is closer to the source material I'm following, and I think it
makes more sense with regards to ray tracing. Ray tracing will involve
floating point mixture and what not, and it makes more sense to work
with this kind of color first.

This adapts some methods in Image to take Into<RGBA>, which if flexible
enough to accmodate things without issue.