Remove existing stuff
It's easier, as we move on to physically based rendering, to remove the
ray tracing code we have now, and start from scratch more or less.

"You should be using a new repo" you despair! Sure. But I don't feel
like it. "Manta" is a pretty good name for a ray tracer, and my intent
was always to make it useable as a standalone program, taking input from
files.

Since I've yet to achieve the goals I've envisioned, and want to have a
more grounded approach in theory, I think it's best to reuse the same
thing.